site/src/backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Jinja2==2.11.2
psycopg2-binary==2.8.5
python-dotenv==0.14.0
SQLAlchemy==1.3.18